+++ sarpyx/cli/__init__.py
@@ -1,27 +1,5 @@
-"""
-CLI module for sarpyx package.
+"""CLI package exposing the WorldSAR entry point only."""
 
-This module provides command-line interfaces for various sarpyx functionalities,
-including ship detection, SAR processing, and other utilities.
-"""
+from .worldsar import main
 
-# Keep imports lazy to reduce CLI startup time.
-__all__ = [
-    'cli_main',
-    'shipdet_main',
-    'utils',
-]
-
-def cli_main(*args, **kwargs):
-    from .main import main
-    return main(*args, **kwargs)
-
-def shipdet_main(*args, **kwargs):
-    from .shipdet import main
-    return main(*args, **kwargs)
-
-def utils():
-    from . import utils as _utils
-    return _utils
-
-__version__ = '0.1.5'
+__all__ = ['main']
