--- sarpyx/cli/algorithms/__init__.py
@@ -1,18 +1,0 @@
-"""Algorithms module for SAR data processing.
-
-This module contains various algorithms for Synthetic Aperture Radar (SAR)
-data processing and analysis.
-"""
-
-__version__ = '0.1.5'
-__author__ = 'SAR Processing Team'
-
-# Import main algorithm modules here when they are created
-# Example:
-# from .calibration import *
-# from .filtering import *
-# from .geocoding import *
-
-__all__ = [
-    # Add algorithm function/class names here as they are implemented
-]
